Change to int64

+++ src/main.h
@@ -27,28 +27,28 @@ class CInv;
 class CRequestTracker;
 class CNode;
 
-static const int MAX_MAGI_POW_HEIGHT = 25000000;
-static const int PRM_MAGI_POW_HEIGHT = 80000;
-static const int PRM_MAGI_POW_HEIGHT_V2 = 50000; // re-cal PoW-I end block
-static const int END_MAGI_POW_HEIGHT = 500000;
-static const int END_MAGI_POW_HEIGHT_V2 = 5000000; // PoW-II aims to issue 12 mil and more than 10 years
+static const int64 MAX_MAGI_POW_HEIGHT = 25000000;
+static const int64 PRM_MAGI_POW_HEIGHT = 80000;
+static const int64 PRM_MAGI_POW_HEIGHT_V2 = 50000; // re-cal PoW-I end block
+static const int64 END_MAGI_POW_HEIGHT = 500000;
+static const int64 END_MAGI_POW_HEIGHT_V2 = 5000000; // PoW-II aims to issue 12 mil and more than 10 years
 
-static const int BLOCK_REWARD_ADJT = 2700;
-static const int BLOCK_REWARD_ADJT_M7M_V2 = 32750;
+static const int64 BLOCK_REWARD_ADJT = 2700;
+static const int64 BLOCK_REWARD_ADJT_M7M_V2 = 32750;
 
 static const unsigned int MAX_BLOCK_SIZE = 1000000;
 static const unsigned int MAX_BLOCK_SIZE_GEN = MAX_BLOCK_SIZE/2;
 static const unsigned int MAX_BLOCK_SIGOPS = MAX_BLOCK_SIZE/50;
 static const unsigned int MAX_ORPHAN_TRANSACTIONS = MAX_BLOCK_SIZE/100;
 static const unsigned int MAX_INV_SZ = 50000;
-static const int MIN_TX_FEE = .0001 * COIN;
-static const int MIN_RELAY_TX_FEE = MIN_TX_FEE;
+static const int64 MIN_TX_FEE = .0001 * COIN;
+static const int64 MIN_RELAY_TX_FEE = MIN_TX_FEE;
 //static const int64 MAX_MONEY_POW_PRM = 10000000 * COIN;	// 10 mil; 5.5 mil in 1st magipow
 //static const int64 MAX_MONEY_POW_END = 15000000 * COIN;	// 15 mil; 5 mil in 2nd magipow
 static const double MAX_MAGI_PROOF_OF_STAKE = 0.05;		// dynamic annual interest, max 5%
 static const double MAX_MAGI_BALANCE_in_STAKE = 0.15;		// balance/money supply, max 15%
-static const int MAX_MONEY_STAKE_REF = 5000000 * COIN;	// 5 mil
-static const int MAX_MONEY_STAKE_REF_V2 = 500000 * COIN;	// 0.5 mil
+static const int64 MAX_MONEY_STAKE_REF = 5000000 * COIN;	// 5 mil
+static const int64 MAX_MONEY_STAKE_REF_V2 = 500000 * COIN;	// 0.5 mil
 
 static const MIN_TXOUT_AMOUNT = MIN_TX_FEE;
